tui: do some gymnastics to circumvent the Sized requirement on `tui::widgets::Widget::render`
We want to get around this requirement so that we can call render on our (?Sized) trait objects.
This lets us do so by essentially by wrapping a reference to the trait object in a new type that implements `TuiWidget`.

I'm fairly certain that the requirement doesn't actually even need to be there since the widget is never passed anywhere by value.

See `Widget::render` and `Frame::render`:
  - https://docs.rs/tui/0.8.0/src/tui/widgets/mod.rs.html#67-87
  - https://docs.rs/tui/0.8.0/src/tui/terminal.rs.html#45-50