tui: switch to allowing mouse events to 'focus' widgets that don't accept it
I think it's actually less surprising behavior for users? But mainly it's easier to implement. In the event that widgets don't accept the focus on a mouse press we need to bubble that information up the tree and that wasn't happening properly in the impl that this commit deletes. Probably an ordering thing.

If there's a need/desire for this behavior we can revisit it.