tui: rejecting unfocusables with mouse based focus _almost_ works
I think it's not trivial to get it to actually work since we'd need a way to feed the actual mous event below (simple focus/not focused things like we do now are not sufficient since the child doesn't know which thing within them the focus event was for).

So, we'll just say that you, as a widget author, must not show borders if your widget can't be focused, I guess.