tui: update the Widget trait to provide a TuiData immutable reference for the draw/render functions
This mirrors the update function. But more importantly, it's totally reasonable for widgets to lean on TuiData to store their drawable state (rather than duplicating it themselves). This was the idea behind providing a Control impl reference in draw/render and there's no reason why TuiData is different in this regard.