tui: add some bad hacks for other program sources on wasm
We can't fetch programs in a blocking manner on WASM so our options are either to fetch the sources _once_ at the start and convert them to immediate sources or to make everything async so that we can fetch things in load_button, from deep within the call stack.

Right now we do the former, but this is obviously not ideal.

I think the latter is basically a non-option because it'd involve re-writing basically everything to be `async` (getting `draw` and `render` on the `Widget` trait to be async also brings up some fun issues — though, `#[async_trait]` traits _are_ still object safe so it's not completely impossible).

So, for now we're stuck with this kludge.

As the contents of this commit mention, we can _potentially_ use WebWorkers and atomics to get around this but it won't be pretty and might not work.